refactor(bench): extract runners module from bench_serving
Move the request/matrix/curve benchmark drivers into runners.rs:
measure_timings, build_request_metrics/iterations, run_info, the three
bench_* drivers, render_text, emit_report, run_command, plus normalize_sizes
and validate_run_args (and the DEFAULT_* prompt consts they use). All are
pub(crate); runners glob-imports its sibling modules.

The cfg(test, deepseek-v2-lite) module's now-local Duration/SamplingParams
imports moved into the test module (they were only used there); trimmed the
profiler/rng imports that left main.rs with measure_timings.

Pure code motion, no behavior change. Verified: default build clean and all
11 deepseek-v2-lite unit tests pass (only the pre-existing ParallelConfig
kimi-gated unused-import warning remains).
